test: refactored interfaces in e2e tests
