+++ test/task.e2e-spec.ts
@@ -2,23 +2,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { INestApplication, ValidationPipe } from '@nestjs/common';
 import * as request from 'supertest';
 import { AppModule } from '../src/app.module';
-import { Types } from 'mongoose';
-
-interface Category {}
-
-interface Task {
-  _id: Types.ObjectId;
-  title: string;
-  description: string;
-  categories: Category[];
-  isCompleted: boolean;
-  dateOfCompletion: null | Date;
-  links: Array<string>;
-  deadline: null | Date;
-  userId: Types.ObjectId;
-  createdAt: Date;
-  updatedAt: Date;
-}
+import { Task } from '../test/interfaces';
 
 describe('Task ontroller (e2e)', () => {
   let app: INestApplication;
